pyproject.toml
[project]
name = "tour-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Tiffany Subin Kim",email = "sbtiffanykim@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2.4,<6.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "pytest-django (>=4.11.1,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["test_*.py", "*_test.py", "testing/python/*.py"]
